feat: dedup all ids on migrations and keep consistent default assets
